+++ rhealpixdggs/projection_wrapper.py
@@ -3,11 +3,6 @@ This Python 3.11 module implements a wrapper for map projections.
 
 - Alexander Raichev (AR), 2013-01-25: Refactored code from release 0.3.
 
-CHANGELOG:
-
-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors
-- RG, 2020-09-08: Issue #6 Added optional region="none" arg to all projection calls
-
 NOTE:
 
 All lengths are measured in meters and all angles are measured in radians