+++ CHANGES.rst
@@ -1,7 +1,46 @@
-=== 0.5.1 ===
-- v0.5.1, 2013-07-29: Rounded floats in doctests.
+0.5.3
+^^^^^
+- v0.5.3: 0.5.1 with:
+-- tests that pass
+-- Sage removed
+-- Test fixes where just relating to rounding errors.
 
-=== 0.5 ===
+- dggs.py:
+-- RG, 2020-09-08: Issue #6 Added optional region="none" arg to rhealpix projection calls, and
+                            forced region to cell.region() in cell.vertex() and cell.boundary()
+-- RG, 2020-07-31: Issue #5 Moved plot_cells to GRS2013 to remove sage dependence
+-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors
+
+- ellipsoids.py:
+-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors
+
+- pj_healpix.py:
+-- RG, 2020-09-08: Issue #6 In in_healpix_image added +-eps to the extreme corner vertices
+                            added calling function abbrev to error statements                            
+-- RG, 2020-07-31: Issue #5 Moved healpix_diagram to GRS2013 to remove sage dependence
+-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors
+
+- pj_rhealpix.py:
+-- RG, 2020-09-08: Issue #6 Added optional region="none" arg to all projection calls, and
+                            used it to enforce region behaviour in calls to combine_triangles
+                            added calling function abbrev to error statements      
+-- RG, 2020-07-31: Issue #5 Moved rhealpix_diagram to GRS2013 to remove sage dependence
+-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors
+
+- projection_wrapper.py:
+-- RG, 2020-09-08: Issue #6 Added optional region="none" arg to all projection calls
+-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors
+
+- utils.py:
+-- Robert Gibb (RG), 2020-07-13: Issue #1 Multiple tests fail due to rounding errors                                                  
+
+0.5.1
+^^^^^
+- v0.5.1, 2013-07-29: Port to Python 3.7
+- Rounded floats in doctests.
+
+0.5
+^^^
 - v0.5, 2013-07-26: Port to Python 3.3.
 
 - dggs.py:
@@ -14,3 +53,7 @@
 
 - pj_healpix.py:
 -- AR, 2013-03-05: In in_healpix_image() increased eps to 1e-10 to decrease out-of-bounds errors i was getting when drawing figures.
+
+Older releases
+^^^^^^^^^^^^^^
+For releases 0.3, 0.4 & 0.5, see the ``_old`` folder within the `0.5.1 release <https://github.com/manaakiwhenua/rhealpixdggs-py/tree/0.5.1/_old>`_ on GitHub